+++ motion-prime/elements/view_with_section.rb
@@ -1,0 +1,7 @@
+module MotionPrime
+  class ViewWithSectionElement < BaseElement
+    def view_class
+      "MPViewWithSection"
+    end
+  end
+end